+++ test/core/config_test.py
@@ -6,12 +6,21 @@ from sqlfluff.core.config import ConfigLoader, nested_combine, dict_diff
 from sqlfluff.core import Linter, FluffConfig
 
 from pathlib import Path
+from unittest.mock import patch, call
+import appdirs
+import pytest
 
 
 config_a = {
     "core": {"testing_val": "foobar", "testing_int": 4},
     "bar": {"foo": "barbar"},
 }
+
+
+@pytest.fixture
+def mock_xdg_home(monkeypatch):
+    """Sets the XDG_CONFIG_HOME variable."""
+    monkeypatch.setenv("XDG_CONFIG_HOME", "~/.config/my/special/path")
 
 
 def test__config__nested_combine():
@@ -63,7 +72,27 @@ def test__config__load_nested():
         )
     )
     assert cfg == {
-        "core": {"testing_val": "foobar", "testing_int": 6, "testing_bar": 7.698},
+        "core": {"testing_val": "foobar", "testing_int": 1, "testing_bar": 7.698},
+        "bar": {"foo": "foobar"},
+        "fnarr": {"fnarr": {"foo": "foobar"}},
+    }
+
+
+def test__config__load_toml():
+    """Test loading config from a pyproject.toml file."""
+    c = ConfigLoader()
+    cfg = c.load_default_config_file(
+        os.path.join("test", "fixtures", "config", "toml"),
+        "pyproject.toml",
+    )
+    assert cfg == {
+        "core": {
+            "testing_int": 5,
+            "testing_bar": 7.698,
+            "testing_bool": False,
+            "testing_arr": ["a", "b", "c"],
+            "testing_inline_table": {"x": 1},
+        },
         "bar": {"foo": "foobar"},
         "fnarr": {"fnarr": {"foo": "foobar"}},
     }
@@ -119,3 +148,36 @@ def test__config__nested_config_tests():
             assert ("L003", 1, 4) in violations[k]
             assert "L002" not in [c[0] for c in violations[k]]
             assert "L009" not in [c[0] for c in violations[k]]
+
+
+@patch("os.path.exists")
+@patch("os.listdir")
+def test__config__load_user_appdir_config(
+    mock_listdir, mock_path_exists, mock_xdg_home
+):
+    """Test loading config from user appdir."""
+    xdg_config_path = os.environ.get("XDG_CONFIG_HOME") + "/sqlfluff"
+
+    def path_exists(x):
+        if x == os.path.expanduser("~/.config/sqlfluff"):
+            return False
+        if x == xdg_config_path:
+            return False
+        else:
+            return True
+
+    mock_path_exists.side_effect = path_exists
+
+    c = ConfigLoader()
+
+    with patch.object(appdirs, attribute="system", new="darwin"):
+        resolved_path = c._get_user_config_dir_path()
+        c.load_user_appdir_config()
+    assert resolved_path == os.path.expanduser("~/Library/Application Support/sqlfluff")
+
+    mock_path_exists.assert_has_calls(
+        [
+            call(xdg_config_path),
+            call(os.path.expanduser("~/Library/Application Support/sqlfluff")),
+        ]
+    )