+++ CHANGELOG.md
@@ -6,21 +6,44 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- L034 now ignores select statements which contain macros.
+- L034 now ignores select statements part of a set expression, most commonly a union.
+- Respect XDG base dirs on Mac OS ([#889](https://github.com/sqlfluff/sqlfluff/issues/889)).
+- Fix bug [#1082](https://github.com/sqlfluff/sqlfluff/issues/1082), adding
+  support for BigQuery `select as struct '1' as bb, 2 as aa` syntax
 
 Contributors:
 - [@GitHub-Username](Link to GitHub profile) ([#PR-Number](Link to PR))
-
+- [@bolajiwahab](https://github.com/bolajiwahab) ([#1063])(https://github.com/sqlfluff/sqlfluff/pull/1063)
+## [0.6.1] - 2021-06-01
+### Changed
+- Added support for additional delimiters by creating a new DelimiterSegment in the 
+  ANSI dialect which defaults to the semicolon, but allows it to be more intuitive 
+  when overriden in a specific child dialect (mysql) [#901](https://github.com/sqlfluff/sqlfluff/issues/901))
+- Added support for the DELIMITER statement in the mysql dialect [#901](https://github.com/sqlfluff/sqlfluff/issues/901))
+## [0.6.0a2] - 2021-05-27
 ### Changed
 - Better exception handling for the simple parsing API (`sqlfluff.parse`)
   which now raises an exception which holds all potential parsing issues
   and prints nicely with more than one issue.
+- Fix bug [#1037](https://github.com/sqlfluff/sqlfluff/issues/1037), in which fix
+  logging had been sent to stdout when reading data from stdin.
+- Add a little bit of fun on CLI exit 🎉!
+- Disabled models in the dbt templater are now skipped enitrely rather than
+  returning an untemplated file.
+- Add a changelog check to SQLFluff continuous integration.
+- Fix bug [#1083](https://github.com/sqlfluff/sqlfluff/issues/1083), adding
+  support for BigQuery named function arguments, used with functions such as
+  [ST_GEOGFROMGEOJSON()](https://cloud.google.com/bigquery/docs/reference/standard-sql/geography_functions#st_geogfromgeojson)
+- Update documentation links to sqlfluff-online.
 
 ## [0.6.0a1] - 2021-05-15
 ### Added
 - Lint and fix parallelism using `--parallel` CLI argument
 - Fix [1051](https://github.com/sqlfluff/sqlfluff/issues/1051), adding support
   for bitwise operators `&`, `|`, `^`, `<<`, `>>`
-  
+
 ## [0.5.6] - 2021-05-14
 - Bugfix release for an issue in `L016` introduced in `0.5.4`.
 - Fix for `L016` issue where `DISTINCT` keywords were mangled during